Simplified script & added new options available in observation file structure
